Split up MaskedInput stories into separate files (#3042)
* Split up MaskedInput stories into separate files

* Added Date story to MaskedInput

* Fixed issues with MaskedInput Date story
