+++ src/js/components/MaskedInput/stories/Time.js
@@ -1,0 +1,52 @@
+import React, { Component } from 'react';
+import { storiesOf } from '@storybook/react';
+import { Box, Grommet, MaskedInput } from 'grommet';
+import { grommet } from 'grommet/themes';
+
+class TimeMaskedInput extends Component {
+  state = { value: '' };
+
+  onChange = event => {
+    this.setState({ value: event.target.value });
+  };
+
+  render() {
+    const { value } = this.state;
+    return (
+      <Grommet full theme={grommet}>
+        <Box fill align="center" justify="start" pad="large">
+          <Box width="medium">
+            <MaskedInput
+              mask={[
+                {
+                  length: [1, 2],
+                  options: Array.from({ length: 12 }, (v, k) => k + 1),
+                  regexp: /^1[0,1-2]$|^0?[1-9]$|^0$/,
+                  placeholder: 'hh',
+                },
+                { fixed: ':' },
+                {
+                  length: 2,
+                  options: ['00', '15', '30', '45'],
+                  regexp: /^[0-5][0-9]$|^[0-9]$/,
+                  placeholder: 'mm',
+                },
+                { fixed: ' ' },
+                {
+                  length: 2,
+                  options: ['am', 'pm'],
+                  regexp: /^[ap]m$|^[AP]M$|^[aApP]$/,
+                  placeholder: 'ap',
+                },
+              ]}
+              value={value}
+              onChange={this.onChange}
+            />
+          </Box>
+        </Box>
+      </Grommet>
+    );
+  }
+}
+
+storiesOf('MaskedInput', module).add('Time', () => <TimeMaskedInput />);
